Refactor: moves exception from application to domain module.
