+++ application/pom.xml
@@ -10,6 +10,14 @@
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>jusbrasil-process-service-application</artifactId>
+    <dependencies>
+        <dependency>
+            <groupId>io.github.ricardosander</groupId>
+            <artifactId>jusbrasil-process-service-domain</artifactId>
+            <version>1.0-SNAPSHOT</version>
+            <scope>compile</scope>
+        </dependency>
+    </dependencies>
 
     <properties>
         <maven.compiler.source>11</maven.compiler.source>
