+++ domain/src/main/java/io/github/ricardosander/jusbrasilprocessservice/domain/InvalidUniqueProcessNumberingException.java
@@ -1,6 +1,6 @@
-package io.github.ricardosander.jusbrasilprocessservice.application;
+package io.github.ricardosander.jusbrasilprocessservice.domain;
 
-class InvalidUniqueProcessNumberingException extends RuntimeException {
+public class InvalidUniqueProcessNumberingException extends RuntimeException {
 
     public InvalidUniqueProcessNumberingException(String uniqueProcessNumbering) {
         super(String.format("%s is an invalid Unique Process Numbering.", uniqueProcessNumbering));
