+++ domain/src/main/java/io/github/ricardosander/jusbrasilprocessservice/domain/NotSupportedUniqueProcessNumberingException.java
@@ -1,6 +1,6 @@
-package io.github.ricardosander.jusbrasilprocessservice.application;
+package io.github.ricardosander.jusbrasilprocessservice.domain;
 
-class NotSupportedUniqueProcessNumberingException extends RuntimeException {
+public class NotSupportedUniqueProcessNumberingException extends RuntimeException {
 
     public NotSupportedUniqueProcessNumberingException(String uniqueProcessNumbering) {
         super(String.format("%s is not supported.", uniqueProcessNumbering));
